Erased Temporary CardView
